requirements.txt
apache-airflow==2.9.1
kafka-python
redis
requests
pandas
